+++ src/ch/fhnw/oop2/model/MInventoryStorage.java
@@ -28,7 +28,7 @@ public class MInventoryStorage extends MInventoryObject {
      * @param image or symbol to use. */
     public MInventoryStorage(int id, String name, String description, CustomImage image, double weight,
                              Color color, double[] dimensions,
-                             double stateOfDecay, Type type, String distinctAttribute) {
+                             int stateOfDecay, Type type, String distinctAttribute) {
 
         super(id, name, description, image, weight, color, dimensions, stateOfDecay, type, distinctAttribute);
         this.objectIds = FXCollections.observableArrayList();
